changed balance field type to decimal, added prettier rendering to admin

+++ accounts/models.py
@@ -2,4 +2,4 @@ from django.db import models
 
 class Account(models.Model):
     name = models.CharField(max_length=50)
-    balance = models.IntegerField(default=0)
+    balance = models.DecimalField(default=0, max_digits=8, decimal_places=2)
